Adjusts the listen_stop program

+++ listen_stop.py
@@ -1,6 +1,15 @@
 import os
 
-pids = os.popen("fuser 9000/tcp").read().replace("  "," ").split(" ")
-pids.remove("")
-for i in pids:
-    os.popen('kill ' + i)
+def killall():    
+    pids = os.popen("fuser 9000/tcp").read().replace("  "," ").split(" ")
+    for i in pids:
+        if i:
+            os.popen('kill ' + i)
+
+os.popen("echo listen_stop activated and will shut program down when p is pressed. | nc 172.20.10.9 9001")
+
+while True:
+    shutdown = input()
+    if shutdown == "p":
+        killall()
+        break
